Lots of work.

+++ src/jtvlc_manager.py
@@ -28,28 +28,31 @@ import vlc_manager
 stream_username = ""
 stream_key = ""
 
+jtvlc_started = False
 jtvlc_process = None
 
 def run_jtvlc(username, key):
-    global jtvlc_process
+    global jtvlc_process, jtvlc_started
     """ Creates a JTVLC instance with the SDP file created by VLC. """
     jtvlc_path = os.path.join(os.getcwd(), "jtvlc-lin-0.41", "jtvlc")
     jtvlc_args = [jtvlc_path,
                   username,
                   key,
                   vlc_manager.sdp_path]
-    jtvlc_process = subprocess.Popen(jtvlc_args)
+    jtvlc_process = subprocess.Popen(jtvlc_args, stderr=subprocess.STDOUT)#, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, close_fds=True)
+    jtvlc_started = True
     print "Created JTVLC with PID %d!" % jtvlc_process.pid
 
 # DOESN'T WORK PLS FIX
 def stop_jtvlc():
     """ Stops a JTVLC instance. """
-    global jtvlc_process
-    if jtvlc_process is not None:
+    global jtvlc_process, jtvlc_started
+    if jtvlc_process is not None and jtvlc_process.poll() is None:
         print "Killing JTVLC with PID %d..." % jtvlc_process.pid
+        jtvlc_started = False
         jtvlc_process.terminate()
         jtvlc_process.wait()
-        jtvlc_process = None
+        #jtvlc_process = None
     else:
         print "Can't kill JTVLC- it isn't running!"
 
@@ -57,9 +60,13 @@ def set_credentials(username, stream_key):
     stream_username = username
     stream_key = stream_key
 
-def is_running():
+def is_process_running():
     global jtvlc_process
     if jtvlc_process is None:
         return False
     else:
-        return jtvlc_process.poll() != None
+        return jtvlc_process.poll() is None
+
+def get_output():
+    global jtvlc_process
+    return jtvlc_process.communicate()[0]
